code/flask/requirements.txt
Flask==2.3.0
Werkzeug==3.0.1
boto3==1.26.109
python-dotenv==1.0.1
Pillow==8.3.2